+++ test/controllers/indexCtrl.js
@@ -1,0 +1,3 @@
+app.controller('indexCtrl', function (ctx, next) {
+  ctx.body = 'This is index page'
+})