+++ test/routes/index.js
@@ -1,0 +1,5 @@
+app.route({
+  method: 'GET',
+  path: '/',
+  controller: 'indexCtrl'
+})